+++ internal/scoreonescriptspan.cc
@@ -184,7 +184,7 @@ void GetTextSpanOffsets(const ScoringHitBuffer* hitbuffer,
 
 
 int DiffScore(const CLD2TableSummary* obj, int indirect,
-              uint16 lang1, uint16 lang2) {
+              uint8 lang1, uint8 lang2) {
   if (indirect < static_cast<int>(obj->kCLDTableSizeOne)) {
     // Up to three languages at indirect
     uint32 langprob = obj->kCLDTableInd[indirect];
@@ -669,7 +669,7 @@ void DumpSummaryBuffer(FILE* df, const SummaryBuffer* summarybuffer) {
 int BetterBoundary(const char* text,
                    ScoringHitBuffer* hitbuffer,
                    ScoringContext* scoringcontext,
-                   uint16 pslang0, uint16 pslang1,
+                   uint8 pslang0, uint8 pslang1,
                    int linear0, int linear1, int linear2) {
   // Degenerate case, no change
   if ((linear2 - linear0) <= 8) {return linear1;}
